requirements.txt
pytest==8.2.1
requests==2.32.3
python-dotenv==1.0.1
jsonschema==4.23.0
allure-pytest==2.13.5
